# Remove live and current indexes