--- a/db/migrate/20190125181528_remove_versioned_editions_current_and_live_indexes.rb
+++ b/db/migrate/20190125181528_remove_versioned_editions_current_and_live_indexes.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+class RemoveVersionedEditionsCurrentAndLiveIndexes < ActiveRecord::Migration[5.2]
+  def change
+    remove_index :versioned_editions, :current
+    remove_index :versioned_editions, :live
+  end
+end